requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
passlib[bcrypt]
python-dotenv
python-multipart
